+++ src/components/02-molecules/Search/Search.scss
@@ -1,0 +1,3 @@
+.m-search {
+  
+}
