+++ src/components/03-organisms/Header/Header.scss
@@ -1,3 +1,5 @@
+@import '../../../styles/style.scss';
+
 .o-header {
   background: $header-background;
   color: $main-text;
