+++ src/components/03-organisms/Navbar/Navbar.scss
@@ -1,0 +1,7 @@
+.o-navbar {
+  // display: inline-block;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
