+++ src/components/util/Layout.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Helmet from 'react-helmet'
 import { StaticQuery, graphql } from "gatsby"
 
-import Navbar from '../organisms/Navbar'
+import Header from '../03-organisms/Header/Header'
 import '../../styles/style.scss'
 
 const TemplateWrapper = ({ children }) => (
@@ -36,7 +36,7 @@ const TemplateWrapper = ({ children }) => (
           <meta property="og:url" content="/" />
           <meta property="og:image" content="/img/og-image.jpg" />
         </Helmet>
-        <Navbar />
+        <Header />
         <div>{children}</div>
       </div>
     )}
