--- src/styles/7-organisms/navbar.scss
@@ -1,4 +1,0 @@
-.o-navbar {
-  // display: inline-block;
-  width: 100%;
-}
