+++ src/styles/style.scss
@@ -10,8 +10,6 @@
 
 @import '4-elements/elements';
 
-@import '7-organisms/header';
-@import '7-organisms/navbar';
 @import '7-organisms/article';
 
 @import '8-pages/pages';
